Bug/deleting nodes sync

+++ src/renderer/ParserTreeInterpreter.ts
@@ -10,7 +10,7 @@ const log = rendererLogger('interpreter');
  */
 interface RendererGraphNode {
     graphNode: GraphNode | undefined;
-    children: RendererGraphNode[];
+    children: Array<RendererGraphNode | undefined>;
 }
 
 /**
@@ -62,7 +62,10 @@ export class ParserTreeInterpreter {
         return curNode;
     }
 
-    private deleteNode(node: RendererGraphNode) {
+    private deleteNode(node: RendererGraphNode | undefined) {
+        if (node === undefined) {
+            throw new Error("Deleting undefined node.");
+        }
         node.graphNode?.remove();
         node.children.forEach(child => {
             this.deleteNode(child);
@@ -111,10 +114,14 @@ export class ParserTreeInterpreter {
                         log.trace('Deleted child');
                         const deletionIndex = Number.parseInt(key.split(/_(\d+)/)[1]);
                         // deleting existing array element
-                        const deletedNode = definedCurNode.children.splice(deletionIndex, 1)[0];
+                        const deletedNode = definedCurNode.children[deletionIndex];
+                        definedCurNode.children[deletionIndex] = undefined;
                         this.deleteNode(deletedNode);
                     }
                 });
+
+                // clean up deleted children
+                definedCurNode.children = definedCurNode.children.filter(child => child !== undefined);
             }
 
         } else {
